+++ services/workflows-service/src/business-report/business-report.controller.internal.ts
@@ -1,6 +1,15 @@
 import * as common from '@nestjs/common';
-import { BadRequestException, Body, Param, Query } from '@nestjs/common';
+import {
+  BadRequestException,
+  Body,
+  Param,
+  Query,
+  Res,
+  UploadedFile,
+  UseInterceptors,
+} from '@nestjs/common';
 import * as swagger from '@nestjs/swagger';
+import { ApiBearerAuth, ApiConsumes } from '@nestjs/swagger';
 import type { InputJsonValue, TProjectId } from '@/types';
 import { AppLoggerService } from '@/common/app-logger/app-logger.service';
 import * as errors from '@/errors';
@@ -29,7 +38,15 @@ import { BusinessReportHookBodyDto } from '@/business-report/dtos/business-repor
 import { BusinessReportHookSearchQueryParamsDto } from '@/business-report/dtos/business-report-hook-search-query-params.dto';
 import { QueryMode } from '@/common/query-filters/query-mode';
 import { isNumber } from 'lodash';
+import { FileInterceptor } from '@nestjs/platform-express';
+import { getDiskStorage } from '@/storage/get-file-storage-manager';
+import { fileFilter } from '@/storage/file-filter';
+import { RemoveTempFileInterceptor } from '@/common/interceptors/remove-temp-file.interceptor';
+import { CreateBusinessReportBatchBodyDto } from '@/business-report/dto/create-business-report-batch-body.dto';
+import type { Response } from 'express';
 
+@ApiBearerAuth()
+@swagger.ApiTags('Business Reports')
 @common.Controller('internal/business-reports')
 @swagger.ApiExcludeController()
 export class BusinessReportControllerInternal {
@@ -53,22 +70,18 @@ export class BusinessReportControllerInternal {
       merchantName,
       businessCorrelationId,
       reportType,
+      workflowVersion,
     }: CreateBusinessReportDto,
     @CurrentProject() currentProjectId: TProjectId,
   ) {
-    const customer = await this.customerService.getByProjectId(currentProjectId);
+    const {
+      id: customerId,
+      displayName: customerName,
+      config,
+    } = await this.customerService.getByProjectId(currentProjectId);
 
-    const { maxBusinessReports, withQualityControl } = customer.config || {};
-
-    if (isNumber(maxBusinessReports) && maxBusinessReports > 0) {
-      const businessReportsCount = await this.businessReportService.count({}, [currentProjectId]);
-
-      if (businessReportsCount >= maxBusinessReports) {
-        throw new BadRequestException(
-          `You have reached the maximum number of business reports allowed (${maxBusinessReports}).`,
-        );
-      }
-    }
+    const { maxBusinessReports, withQualityControl } = config || {};
+    await this.checkBusinessReportsLimit(maxBusinessReports, currentProjectId);
 
     let business: Pick<Business, 'id' | 'correlationId'> | undefined;
     const merchantNameWithDefault = merchantName || 'Not detected';
@@ -121,11 +134,12 @@ export class BusinessReportControllerInternal {
         countryCode,
         parentCompanyName: merchantName,
         reportType,
+        workflowVersion,
         withQualityControl,
         callbackUrl: `${env.APP_API_URL}/api/v1/internal/business-reports/hook?businessId=${business.id}&businessReportId=${businessReport.id}`,
         metadata: {
-          customerId: customer.id,
-          customerName: customer.displayName,
+          customerId,
+          customerName,
           workflowRuntimeDataId: null,
         },
       },
@@ -249,11 +263,12 @@ export class BusinessReportControllerInternal {
   @common.UsePipes(new ZodValidationPipe(ListBusinessReportsSchema, 'query'))
   async listBusinessReports(
     @CurrentProject() currentProjectId: TProjectId,
-    @Query() { businessId, page, search, type, orderBy }: ListBusinessReportsDto,
+    @Query() { businessId, batchId, page, search, type, orderBy }: ListBusinessReportsDto,
   ) {
     const args = {
       where: {
         businessId,
+        batchId,
         ...(type ? { type } : {}),
         ...(search
           ? {
@@ -331,5 +346,58 @@ export class BusinessReportControllerInternal {
         },
       },
     });
+  }
+
+  @common.Post('/upload-batch')
+  @swagger.ApiForbiddenResponse({ type: errors.ForbiddenException })
+  @ApiConsumes('multipart/form-data')
+  @UseInterceptors(
+    FileInterceptor('file', {
+      storage: getDiskStorage(),
+      fileFilter,
+    }),
+    RemoveTempFileInterceptor,
+  )
+  async createBusinessReportBatch(
+    @UploadedFile() file: Express.Multer.File,
+    @Body() { type, workflowVersion }: CreateBusinessReportBatchBodyDto,
+    @Res() res: Response,
+    @CurrentProject() currentProjectId: TProjectId,
+  ) {
+    const { config } = await this.customerService.getByProjectId(currentProjectId);
+
+    const { maxBusinessReports, withQualityControl } = config || {};
+    await this.checkBusinessReportsLimit(maxBusinessReports, currentProjectId);
+
+    const result = await this.businessReportService.processBatchFile({
+      type,
+      workflowVersion,
+      currentProjectId,
+      maxBusinessReports,
+      merchantSheet: file,
+      projectId: currentProjectId,
+      withQualityControl: typeof withQualityControl === 'boolean' ? withQualityControl : false,
+    });
+
+    res.status(201);
+    res.setHeader('content-type', 'application/json');
+    res.send(result);
+  }
+
+  private async checkBusinessReportsLimit(
+    maxBusinessReports: number | undefined,
+    currentProjectId: string,
+  ) {
+    if (!isNumber(maxBusinessReports) || maxBusinessReports <= 0) {
+      return;
+    }
+
+    const businessReportsCount = await this.businessReportService.count({}, [currentProjectId]);
+
+    if (businessReportsCount >= maxBusinessReports) {
+      throw new BadRequestException(
+        `You have reached the maximum number of business reports allowed (${maxBusinessReports}).`,
+      );
+    }
   }
 }