+++ packages/common/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @ballerine/common
 
+## 0.9.28
+
+### Patch Changes
+
+- Version bump
+
+## 0.9.27
+
+### Patch Changes
+
+- bump
+
+## 0.9.26
+
+### Patch Changes
+
+- pushing fixes
+
+## 0.9.25
+
+### Patch Changes
+
+- version bump
+
+## 0.9.24
+
+### Patch Changes
+
+- merchant screening changes
+
+## 0.9.23
+
+### Patch Changes
+
+- added merchant screening to swagger
+
 ## 0.9.22
 
 ### Patch Changes