+++ packages/config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ballerine/config
 
+## 1.1.13
+
+### Patch Changes
+
+- Version bump
+
+## 1.1.12
+
+### Patch Changes
+
+- bump
+
 ## 1.1.11
 
 ### Patch Changes