+++ sdks/workflow-node-sdk/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @ballerine/workflow-node-sdk
 
+## 0.6.39
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.39
+
+## 0.6.38
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.38
+
+## 0.6.37
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.37
+
+## 0.6.36
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.36
+
+## 0.6.35
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.35
+
+## 0.6.34
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.6.34
+
 ## 0.6.33
 
 ### Patch Changes